% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qcarchive}
\alias{qcarchive}
\title{This is data to be included in my package}
\format{
An object of class \code{data.frame} with 43532 rows and 163 columns.
}
\source{
\href{https://phenome.jax.org/projects/Moore1b}{QTL Archive}
}
\usage{
data(qcarchive)
}
\description{
This is data to be included in my package
}
\references{
Moore et al. (2013) Genetics 195:1077-1086
(\href{https://www.ncbi.nlm.nih.gov/pubmed/23979570}{PubMed})
}
\keyword{datasets}
